Auth/backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
casbin==1.34.0
casbin-sqlalchemy-adapter==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.0
ragflow-sdk>=0.12.0
